2.0.6 button upgrade, fancytools and positioning system fixes
adding buttons on the webUI

the button system is modified and the gpio buttons are directly linked to the hardware active display. The webUI buttons screen attribution is selected by the user,=.

fixing the FancyTools links
the mapping system is different, and instead of menu its replaced with --button -B.
ex:
fancytools -B up

the positioning is fixed and formula can be used on any widget type again.