2.0.7 revert the theme attribution
revert the theme update attribution. Sometimes the toml file is not loaded properly and this is giving an error in the log.

the fix attempt caused a reset of the default theme and rotation in the file, in the case where the config file was not loaded properly.

A single error is better than this reset. Ill search for another fix.